requirements-tests.txt
pytz
qrcode
xhtml2pdf
django-sekizai
